+++ psycopg/psycopg/_struct.py
@@ -6,10 +6,11 @@ Utility functions to deal with binary structs.
 
 import struct
 from typing import Callable, cast, Optional, Tuple
+from typing_extensions import TypeAlias
 
 from .abc import Buffer
 from . import errors as e
-from ._compat import Protocol, TypeAlias
+from ._compat import Protocol
 
 PackInt: TypeAlias = Callable[[int], bytes]
 UnpackInt: TypeAlias = Callable[[Buffer], Tuple[int]]