+++ psycopg/psycopg/types/enum.py
@@ -4,13 +4,13 @@ Adapters for the enum type.
 from enum import Enum
 from typing import Any, Dict, Generic, Optional, Mapping, Sequence
 from typing import Tuple, Type, TypeVar, Union, cast
+from typing_extensions import TypeAlias
 
 from .. import postgres
 from .. import errors as e
 from ..pq import Format
 from ..abc import AdaptContext
 from ..adapt import Buffer, Dumper, Loader
-from .._compat import TypeAlias
 from .._encodings import conn_encoding
 from .._typeinfo import EnumInfo as EnumInfo  # exported here